Refactor float data types
Floats data format causing problems - switch to double

Response204 object removed to more accurately reflect the behavior of the SDK.  the 204 no content resulted in responses which could not be parsed.  The SDK was "faking" a JSON response object - but now it returns void for delete methods that result in 204 response.

Added "debug" level logging for API calls for
- url
- post/put body
- headers
- response
changed TLS warning log to "trace" level for logging.

Updating descriptions of responses and params to improve JavaDocs

Added missing attributes of Invoices
- hasError boolean
- Attachments Array
- isDiscounted boolean
Added missing attributes of Payments
- hasAccount
- hasValidationErrors

Added more Unit Tests - 60 sleep between each one to avoid hitting rate limits when testing against mock API server.

+++ src/main/java/com/xero/models/accounting/AccountsPayable.java
@@ -30,13 +30,13 @@ import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 public class AccountsPayable {
   
   @JsonProperty("Outstanding")
-  private Float outstanding;
+  private Double outstanding;
 
   
   @JsonProperty("Overdue")
-  private Float overdue;
+  private Double overdue;
 
-  public AccountsPayable outstanding(Float outstanding) {
+  public AccountsPayable outstanding(Double outstanding) {
     this.outstanding = outstanding;
     return this;
   }
@@ -46,15 +46,15 @@ public class AccountsPayable {
    * @return outstanding
   **/
   @ApiModelProperty(value = "")
-  public Float getOutstanding() {
+  public Double getOutstanding() {
     return outstanding;
   }
 
-  public void setOutstanding(Float outstanding) {
+  public void setOutstanding(Double outstanding) {
     this.outstanding = outstanding;
   }
 
-  public AccountsPayable overdue(Float overdue) {
+  public AccountsPayable overdue(Double overdue) {
     this.overdue = overdue;
     return this;
   }
@@ -64,11 +64,11 @@ public class AccountsPayable {
    * @return overdue
   **/
   @ApiModelProperty(value = "")
-  public Float getOverdue() {
+  public Double getOverdue() {
     return overdue;
   }
 
-  public void setOverdue(Float overdue) {
+  public void setOverdue(Double overdue) {
     this.overdue = overdue;
   }
