+++ src/main/java/com/xero/models/accounting/Purchase.java
@@ -30,7 +30,7 @@ import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 public class Purchase {
   
   @JsonProperty("UnitPrice")
-  private Float unitPrice;
+  private Double unitPrice;
 
   
   @JsonProperty("AccountCode")
@@ -44,7 +44,7 @@ public class Purchase {
   @JsonProperty("TaxType")
   private String taxType;
 
-  public Purchase unitPrice(Float unitPrice) {
+  public Purchase unitPrice(Double unitPrice) {
     this.unitPrice = unitPrice;
     return this;
   }
@@ -54,11 +54,11 @@ public class Purchase {
    * @return unitPrice
   **/
   @ApiModelProperty(value = "Unit Price of the item. By default UnitPrice is rounded to two decimal places. You can use 4 decimal places by adding the unitdp=4 querystring parameter to your request.")
-  public Float getUnitPrice() {
+  public Double getUnitPrice() {
     return unitPrice;
   }
 
-  public void setUnitPrice(Float unitPrice) {
+  public void setUnitPrice(Double unitPrice) {
     this.unitPrice = unitPrice;
   }
